// src/app/models/weather.model.ts
export interface Weather {
    city: any;
    country: any;
    iconUrl: any;
    temperature: any;
    name: string;
    description: string;
    icon: string;
    temp: number;
    tempMin: number;
    tempMax: number;
    feelsLike: number;
    humidity: number;
    windSpeed: number;
    sunrise: Date;
    sunset: Date;
}